Implement GravatarQuickEditorActivity
To simplify the integration we need to remove the limitation of using the QuickEditor from an activity using the launchMode="singleTask".

The approach is to create an intermediate activity that can assume the complexity and the launch mode required by the QE. The behavior should be as much as possible as it is using the Composable directly.

In this commit, we are losing the "callback" part, so the app can't notice when the QE is closed or when the avatar changes. We'll need to add some ways to inform you about those events.

# Conflicts:
#	gravatar-quickeditor/build.gradle.kts
#	gravatar/build.gradle.kts