Delete avatar logic implemented in AvatarPickerViewModel
The delete avatar logic has been implemented, including the failure case where we should add the remove avatar to the UI again.

We assume the deletion will be completed successfully, so we will immediately remove the avatar from the list. In case of an error, we need to add it again.